chore: move folders out of storage folder
