+++ src/main/java/com/gdg/kkia/chatbot/repository/ChatbotResponseRepository.java
@@ -1,10 +1,8 @@
-package com.gdg.kkia.storage.chatbot.repository;
+package com.gdg.kkia.chatbot.repository;
 
-import com.gdg.kkia.storage.chatbot.entity.ChatbotResponse;
+import com.gdg.kkia.chatbot.entity.ChatbotResponse;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
 
 @Repository
 public interface ChatbotResponseRepository extends JpaRepository<ChatbotResponse, Long> {
