+++ src/main/java/com/gdg/kkia/chatbot/service/ChatbotResponseService.java
@@ -1,13 +1,13 @@
-package com.gdg.kkia.storage.chatbot.service;
+package com.gdg.kkia.chatbot.service;
 
+import com.gdg.kkia.chatbot.repository.ChatbotResponseRepository;
 import com.gdg.kkia.common.exception.BadRequestException;
 import com.gdg.kkia.common.exception.NotFoundException;
 import com.gdg.kkia.gemini.entity.GeminiRequestType;
 import com.gdg.kkia.gemini.dto.GeminiContent;
 import com.gdg.kkia.member.entity.Member;
 import com.gdg.kkia.member.repository.MemberRepository;
-import com.gdg.kkia.storage.chatbot.entity.ChatbotResponse;
-import com.gdg.kkia.storage.chatbot.repository.ChatbotResponseRepository;
+import com.gdg.kkia.chatbot.entity.ChatbotResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
