+++ src/main/java/com/gdg/kkia/dailyresponse/dto/DailyQuestionRequest.java
@@ -1,4 +1,4 @@
-package com.gdg.kkia.storage.dailyresponse.dto;
+package com.gdg.kkia.dailyresponse.dto;
 
 public record DailyQuestionRequest(
         String question
