+++ src/main/java/com/gdg/kkia/dailyresponse/dto/DailyQuestionResponse.java
@@ -1,4 +1,4 @@
-package com.gdg.kkia.storage.dailyresponse.dto;
+package com.gdg.kkia.dailyresponse.dto;
 
 public record DailyQuestionResponse(
         Long id,
