+++ src/main/java/com/gdg/kkia/dailyresponse/dto/DailyResponseRequest.java
@@ -1,4 +1,4 @@
-package com.gdg.kkia.storage.dailyresponse.dto;
+package com.gdg.kkia.dailyresponse.dto;
 
 public record DailyResponseRequest(
         Long questionId,
