+++ src/main/java/com/gdg/kkia/dailyresponse/dto/DailyResponseResponse.java
@@ -1,4 +1,4 @@
-package com.gdg.kkia.storage.dailyresponse.dto;
+package com.gdg.kkia.dailyresponse.dto;
 
 public record DailyResponseResponse(
         Long id,
