+++ src/main/java/com/gdg/kkia/dailyresponse/repository/DailyQuestionRepository.java
@@ -1,6 +1,6 @@
-package com.gdg.kkia.storage.dailyresponse.repository;
+package com.gdg.kkia.dailyresponse.repository;
 
-import com.gdg.kkia.storage.dailyresponse.entity.DailyQuestion;
+import com.gdg.kkia.dailyresponse.entity.DailyQuestion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DailyQuestionRepository extends JpaRepository<DailyQuestion, Long> {
