+++ src/main/java/com/gdg/kkia/diary/repository/DiaryRepository.java
@@ -1,6 +1,6 @@
-package com.gdg.kkia.storage.diary.repository;
+package com.gdg.kkia.diary.repository;
 
-import com.gdg.kkia.storage.diary.entity.Diary;
+import com.gdg.kkia.diary.entity.Diary;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DiaryRepository extends JpaRepository<Diary, Long> {
